split Mutation file into multiple files
