+++ apps/dh/api-dh/source/DataHub.WebApi/GraphQL/Mutation/AggregatedMeasureDataMutation.cs
@@ -1,0 +1,48 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.Edi.B2CWebApp.Clients.v1;
+using EdiB2CWebAppProcessType = Energinet.DataHub.Edi.B2CWebApp.Clients.v1.ProcessType;
+
+namespace Energinet.DataHub.WebApi.GraphQL.Mutation;
+
+public partial class Mutation
+{
+    public async Task<bool> CreateAggregatedMeasureDataRequestAsync(
+        EdiB2CWebAppProcessType processType,
+        MeteringPointType? meteringPointType,
+        string startDate,
+        string? endDate,
+        string? gridArea,
+        string? energySupplierId,
+        string? balanceResponsibleId,
+        CancellationToken cancellationToken,
+        [Service] IEdiB2CWebAppClient_V1 client)
+    {
+        await client.RequestAggregatedMeasureDataAsync(
+            new RequestAggregatedMeasureDataMarketRequest()
+            {
+                ProcessType = processType,
+                MeteringPointType = meteringPointType,
+                StartDate = startDate,
+                EndDate = endDate,
+                GridArea = gridArea,
+                EnergySupplierId = energySupplierId,
+                BalanceResponsibleId = balanceResponsibleId,
+            },
+            cancellationToken)
+            .ConfigureAwait(false);
+        return true;
+    }
+}
